# docs/swagger.yaml
basePath: /api/v1
definitions:
...
  disaster_alert_backend_internal_dto.MarkAppNotificationReadRequest:
    properties:
      notificationIds:
        items:
          type: string
        type: array
    type: object
...
  disaster_alert_backend_internal_dto.UpdateAlertPreferenceRequest:
    properties:
      conflict:
        description: Conflict alerts.
        example: true
        type: boolean
      criticalAlerts:
        description: CriticalAlerts controls whether the user receives critical/highest
          priority alerts.
        example: true
        type: boolean
      drought:
        description: Drought alerts.
        example: false
        type: boolean
      earthquake:
        description: Earthquake alerts.
        example: false
        type: boolean
      fire:
        description: Fire alerts.
        example: true
        type: boolean
      flood:
        description: Flood alerts.
        example: true
        type: boolean
      galamsey:
        description: Galamsey or illegal mining-related alerts.
        example: false
        type: boolean
      health:
        description: Health alerts such as outbreaks, contamination, or public health
          warnings.
        example: true
        type: boolean
      munitions:
        description: Munitions or explosive-related alerts.
        example: true
        type: boolean
      protests:
        description: Protest or civil unrest alerts.
        example: true
        type: boolean
      robbery:
        description: Robbery or crime-related alerts.
        example: true
        type: boolean
      unverifiedActivity:
        description: Unverified activity alerts.
        example: false
        type: boolean
      weather:
        description: Weather alerts such as storms, heavy rainfall, heat, wind, or
          severe weather.
        example: true
        type: boolean
    type: object
...
info:
  contact: {}
  description: |-
    Disaster Alert API is a mission-control backend for disaster intelligence, community incident reports, SOS escalation, assistance coordination, emergency messaging, public alerts, chats, notifications, and administrator response operations.

    ================================
    QUICK START
    ================================

    Use this Swagger page to test the API safely.

    For mobile/user endpoints:
    1. Register or login using /auth/register or /auth/login.
    2. Copy the token from the login response.
    3. Click Authorize.
    4. Paste it like this: Bearer YOUR_JWT_TOKEN.
    5. Test endpoints such as /reports, /alerts, /sos, /assistance, /notifications, and /chats.

    For admin endpoints:
    1. Enter the Admin API Key under AdminApiKeyAuth.
    2. Create or validate a privilege code.
    3. Enter the full privilege code under PrivilegeCodeAuth.
    4. Test admin endpoints such as /admin/reports, /admin/alerts, /admin/sos, /admin/assistance, and /admin/chats.

    ================================
    AUTHENTICATION GUIDE
    ================================

    1. Mobile/User endpoints use BearerAuth.
    Required header:
    Authorization: Bearer <JWT_TOKEN>

    2. Basic admin management endpoints use AdminApiKeyAuth.
    Required header:
    Sigtrack-Admin-API-Key: <ADMIN_API_KEY>

    3. Privileged admin operation endpoints require BOTH AdminApiKeyAuth and PrivilegeCodeAuth.
...

...

...

...

...

...

...

...

...

...

...

...

...

...

...
paths:
...
  /notifications:
    get:
      description: |-
        Returns in-app notifications for the authenticated user.

        WHEN TO USE THIS ENDPOINT:
        Use this to show the user's notification inbox inside the mobile app.

        AUTH REQUIRED:
        This endpoint requires a valid user JWT token.
        Click Authorize and paste: Bearer YOUR_JWT_TOKEN.

        QUERY PARAMETERS:
        - limit: Optional number of notifications to return. Default is 50.
      parameters:
      - description: Maximum number of notifications to return. Default is 50.
        example: 50
        in: query
        name: limit
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: Notifications fetched successfully.
          schema:
            additionalProperties: true
            type: object
        "401":
          description: Missing, invalid, or expired user token.
          schema:
            additionalProperties: true
            type: object
        "500":
          description: Failed to fetch notifications.
          schema:
            additionalProperties: true
            type: object
      security:
      - BearerAuth: []
      summary: Get my notifications
      tags:
      - User Notifications
  /notifications/read:
    put:
      consumes:
      - application/json
      description: |-
        Marks selected notifications as read for the authenticated user.

        WHEN TO USE THIS ENDPOINT:
        Use this when the user opens one or more notifications and they should no longer appear as unread.

        AUTH REQUIRED:
        This endpoint requires a valid user JWT token.
        Click Authorize and paste: Bearer YOUR_JWT_TOKEN.

        REQUIRED BODY:
        Send notificationIds as a list of notification IDs.

        EXAMPLE REQUEST BODY:
        {
        "notificationIds": [
        "66e19b71c8f2a2b4d1234567"
        ]
        }
      parameters:
      - description: Notification read payload. notificationIds is the list of notification
          IDs to mark as read.
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/disaster_alert_backend_internal_dto.MarkAppNotificationReadRequest'
      produces:
      - application/json
      responses:
        "200":
          description: Notification marked as read.
          schema:
            additionalProperties: true
            type: object
        "400":
          description: Invalid request body or notification could not be marked as
            read.
          schema:
            additionalProperties: true
            type: object
        "401":
          description: Missing, invalid, or expired user token.
          schema:
            additionalProperties: true
            type: object
        "500":
          description: Server error while marking notification as read.
          schema:
            additionalProperties: true
            type: object
      security:
      - BearerAuth: []
      summary: Mark notifications as read
      tags:
      - User Notifications
  /notifications/read-all:
    put:
      description: |-
        Marks all in-app notifications as read for the authenticated user.

        WHEN TO USE THIS ENDPOINT:
        Use this when the user taps a “Mark all as read” button in the mobile app.

        AUTH REQUIRED:
        This endpoint requires a valid user JWT token.
        Click Authorize and paste: Bearer YOUR_JWT_TOKEN.
      produces:
      - application/json
      responses:
        "200":
          description: All notifications marked as read.
          schema:
            additionalProperties: true
            type: object
        "401":
          description: Missing, invalid, or expired user token.
          schema:
            additionalProperties: true
            type: object
        "500":
          description: Failed to mark notifications as read.
          schema:
            additionalProperties: true
            type: object
      security:
      - BearerAuth: []
      summary: Mark all notifications as read
      tags:
      - User Notifications
...
  /notifications/unread-count:
    get:
      description: |-
        Returns the number of unread in-app notifications for the authenticated user.

        WHEN TO USE THIS ENDPOINT:
        Use this to show a badge count on the notification icon in the mobile app.

        AUTH REQUIRED:
        This endpoint requires a valid user JWT token.
        Click Authorize and paste: Bearer YOUR_JWT_TOKEN.
      produces:
      - application/json
      responses:
        "200":
          description: Unread notification count fetched successfully. Response data
            contains count.
          schema:
            additionalProperties: true
            type: object
        "401":
          description: Missing, invalid, or expired user token.
          schema:
            additionalProperties: true
            type: object
        "500":
          description: Failed to fetch unread notification count.
          schema:
            additionalProperties: true
            type: object
      security:
      - BearerAuth: []
      summary: Get unread notification count
      tags:
      - User Notifications
...
  /user/alert-preferences:
    get:
      description: |-
        Returns the authenticated user's alert category preferences.

        WHEN TO USE THIS ENDPOINT:
        Use this when the app opens the user's alert settings/preferences screen.
      produces:
      - application/json
      responses:
        "200":
          description: Alert preferences fetched successfully.
          schema:
            additionalProperties: true
            type: object
        "401":
          description: Missing, invalid, or expired user token.
          schema:
            additionalProperties: true
            type: object
        "500":
          description: Failed to fetch alert preferences.
          schema:
            additionalProperties: true
            type: object
      security:
      - BearerAuth: []
      summary: Get alert preferences
      tags:
      - User Settings
    put:
      consumes:
      - application/json
      description: |-
        Updates the authenticated user's alert category preferences.

        WHEN TO USE THIS ENDPOINT:
        Use this when the user enables or disables alert categories in the app settings.

        IMPORTANT:
        All fields are optional. Send only the categories you want to change.
        Use true to enable a category and false to disable it.

        EXAMPLE REQUEST BODY:
        {
        "fire": true,
        "flood": true,
        "weather": true,
        "health": true,
        "criticalAlerts": true
        }
      parameters:
      - description: Alert preference update payload. All fields are optional.
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/disaster_alert_backend_internal_dto.UpdateAlertPreferenceRequest'
      produces:
      - application/json
      responses:
        "200":
          description: Alert preferences updated successfully.
          schema:
            additionalProperties: true
            type: object
        "400":
          description: Invalid request body.
          schema:
            additionalProperties: true
            type: object
        "401":
          description: Missing, invalid, or expired user token.
          schema:
            additionalProperties: true
            type: object
        "500":
          description: Failed to update alert preferences.
          schema:
            additionalProperties: true
            type: object
      security:
      - BearerAuth: []
      summary: Update alert preferences
      tags:
      - User Settings
...
schemes:
...
securityDefinitions:
  AdminApiKeyAuth:
    in: header
    name: Sigtrack-Admin-API-Key
    type: apiKey
  BearerAuth:
    in: header
    name: Authorization
    type: apiKey
  PrivilegeCodeAuth:
    in: header
    name: X-Privilege-Code
    type: apiKey
swagger: "2.0"